// packages/material/src/editor-ui/mediaComponent/ItemComponent/enum.ts
// 枚举定义
export enum EMixBlendMode {
  Normal = "normal",
  Multiply = "multiply",
  Screen = "screen",
  Overlay = "overlay",
  Darken = "darken",
  Lighten = "lighten",
  ColorDodge = "color-dodge",
  ColorBurn = "color-burn",
  HardLight = "hard-light",
  SoftLight = "soft-light",
  Difference = "difference",
  Exclusion = "exclusion",
  Hue = "hue",
  Saturation = "saturation",
  Color = "color",
  Luminosity = "luminosity"
}

export enum EAnimationSpeed {
  Constant = "constant",
  SlowFastSlow = "slow-fast-slow",
  StartSlow = "start-slow",
  EndSlow = "end-slow"
}

export enum EAnimationType {
  Default = "default",
  Opacity = "opacity",
  Zoom = "zoom",
  Clockwise = "clockwise",
  Counterclockwise = "counterclockwise",
  BackAndForth = "backAndForth",
  UpOrDown = "upOrDown"
}

export enum imgEAnimationType {
  Default = "default",
  Opacity = "opacity",
  Zoom = "zoom",
  Clockwise = "clockwise",
  Counterclockwise = "counterclockwise",
  BackAndForth = "backAndForth",
  UpOrDown = "upOrDown",
  Customize = "customize"
}

export enum EBackgroundType {
  Color = "color",
  Custom = "custom"
}

export enum EBackgroundImageType {
  Adapt = "100% 100%",
  Contain = "contain",
  Cover = "cover"
}

export enum EArrowOption {
  Always = "always",
  Hover = "hover",
  Never = "never"
}

export enum ESeriesLabelOrient {
  Horizontal = "horizontal",
  Vertical = "vertical"
}

export enum EObjectFit {
  None = "none",
  Fill = "fill",
  ScaleDown = "scale-down",
  Cover = "cover",
  Contain = "contain"
}

export enum ETextAlign {
  Center = "center",
  Left = "left",
  Right = "right"
}